refactor : Change drectiry
sample 패키지로 묶음
