+++ app/src/main/java/com/alpaca/app/sample/controller/SampleController.java
@@ -1,4 +1,4 @@
-package com.alpaca.app.controller;
+package com.alpaca.app.sample.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
