+++ app/src/main/resources/mybatis-mapper/HRMapper.xml
@@ -2,7 +2,7 @@
 
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.alpaca.app.mapper.HRMapper">
+<mapper namespace="com.alpaca.app.sample.mapper.HRMapper">
     <select id="getName2" resultType="string">
 
         SELECT
