+++ app/src/main/resources/mybatis-mapper/TimeMapper.xml
@@ -2,7 +2,7 @@
 
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="TimeMapper">
+<mapper namespace="com.alpaca.app.sample.mapper.TimeMapper">
     <select id="getTime2" resultType="string">
         select sysdate from dual
     </select>
